學員作品/葉俐彤_中華職棒球員資料查詢/requirements.txt
matplotlib
matplotlib-inline
Pillow
ttkbootstrap
seaborn